requirements.txt
aiofiles==24.1.0
aiogram==3.15.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.11
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.7.0
apache-libcloud==3.8.0
asyncpg==0.30.0
attrs==24.2.0
bcrypt==4.2.1
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
fasteners==0.19
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
magic-filter==1.0.12
MarkupSafe==3.0.2
multidict==6.1.0
propcache==0.2.0
psycopg2-binary==2.9.10
pydantic==2.9.2
pydantic_core==2.23.4
python-dotenv==1.0.1
python-multipart==0.0.19
requests==2.32.3
sniffio==1.3.1
SQLAlchemy==2.0.36
sqlalchemy-file==0.6.0
starlette==0.41.3
starlette-admin==0.14.1
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.1
yarl==1.18.0
